backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "pytest>=8.3.4",
    "httpx>=0.28.1",
    "openai>=2.14.0",
]
